requirements.txt
# Core Framework
langchain>=0.1.0
langchain-community>=0.0.20
openai>=1.0.0
anthropic>=0.7.0

# Trading & Market Data
alpaca-py>=0.5.0
finnhub-python>=2.4.0
yfinance>=0.2.0
pandas>=2.0.0
numpy>=1.24.0
ta>=0.10.0

# Data Processing
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Database & Storage
redis>=4.5.0
sqlalchemy>=2.0.0
alembic>=1.12.0

# Monitoring & Alerting
prometheus-client>=0.17.0
slack-sdk>=3.21.0
python-telegram-bot>=20.0

# Web and API
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
responses>=0.23.0

# Security
cryptography>=41.0.0
PyJWT>=2.8.0

# Configuration
PyYAML>=6.0

# Vector Store and Embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# Additional utilities
pytz>=2023.3
schedule>=1.2.0
asyncpg>=0.29.0
aiofiles>=23.2.1